+++ tests/Pim/MultiVendorProduct/MultiVendorProductServiceTest.php
@@ -11,6 +11,7 @@ namespace Wizaplace\SDK\Tests\Pim\MultiVendorProduct;
 
 use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
 use Wizaplace\SDK\File\File;
+use Wizaplace\SDK\Image\Image;
 use Wizaplace\SDK\Pagination;
 use Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProduct;
 use Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductFile;
@@ -346,8 +347,36 @@ final class MultiVendorProductServiceTest extends ApiTestCase
 
     public function testAddImageToMultiVendorProduct()
     {
-        $service = $this->buildMultiVendorProductService();
-        $uuid = '0adaf6bc-d362-34be-b72f-42d5aa3b4a4e';
+        $service = $this->buildMultiVendorProductService('admin@wizaplace.com', static::VALID_PASSWORD);
+        $newMvp = new MultiVendorProduct(
+            [
+                'name' => 'New Test MVP T',
+                'code' => 'XXX-xxx-XXY',
+                'supplierReference' => 'AZAZPMPKKD24',
+                'productTemplateType' => 'product',
+                'slug' => 'new-test-mvp-y',
+                'shortDescription' => 'Nihil est qui quibusdam exercitationem',
+                'description' => 'Nihil est qui quibusdam exercitationem consequatur doloribus sit velit. Ut temporibus est qui et molestiae facilis nisi',
+                'seoTitle' => 'New Test MVP',
+                'seoDescription' => 'New Test MVP Nihil est qui quibusdam exercitationem',
+                'seoKeywords' => 'MVP, PRODUCT, TEST',
+                'status' => 'A',
+                'categoryId' => 3,
+                'freeAttributes' => [
+                    'Free attribute multiple' => [
+                        'Num 1',
+                        'Num 2',
+                        51,
+                    ],
+                    'Free attribute simple' => [
+                        'Bla',
+                    ],
+                ],
+                'attributes' => [],
+            ]
+        );
+
+        $uuid = $service->createMultiVendorProduct($newMvp);
 
         $image = $this->mockUploadedFile("favicon.png");
 
@@ -359,6 +388,100 @@ final class MultiVendorProductServiceTest extends ApiTestCase
 
         $this->assertInstanceOf(MultiVendorProduct::class, $multiVendorProduct);
         $this->assertEquals($uuid, $multiVendorProduct->getId());
+    }
+
+    public function testAddImageToMultiVendorProductWithAlt(): void
+    {
+        $service = $this->buildMultiVendorProductService('admin@wizaplace.com', static::VALID_PASSWORD);
+
+        $newMvp = new MultiVendorProduct(
+            [
+                'name' => 'New Test MVP',
+                'code' => 'XXX-xxx-XXX',
+                'supplierReference' => 'AZAZPMPKKD23',
+                'productTemplateType' => 'product',
+                'slug' => 'new-test-mvp',
+                'shortDescription' => 'Nihil est qui quibusdam exercitationem',
+                'description' => 'Nihil est qui quibusdam exercitationem consequatur doloribus sit velit. Ut temporibus est qui et molestiae facilis nisi',
+                'seoTitle' => 'New Test MVP',
+                'seoDescription' => 'New Test MVP Nihil est qui quibusdam exercitationem',
+                'seoKeywords' => 'MVP, PRODUCT, TEST',
+                'status' => 'A',
+                'categoryId' => 3,
+                'freeAttributes' => [
+                    'Free attribute multiple' => [
+                        'Num 1',
+                        'Num 2',
+                        51,
+                    ],
+                    'Free attribute simple' => [
+                        'Bla',
+                    ],
+                ],
+                'attributes' => [],
+            ]
+        );
+
+        $uuid = $service->createMultiVendorProduct($newMvp);
+
+        $image = $this->mockUploadedFile("favicon.png");
+
+        $files = [
+            new MultiVendorProductFile('file', $image->getStream(), $image->getClientFilename()),
+        ];
+
+        $multiVendorProduct = $service->addImageToMultiVendorProduct($uuid, $files, 'alternative text mvp');
+
+        static::assertInstanceOf(MultiVendorProduct::class, $multiVendorProduct);
+        static::assertInstanceOf(Image::class, $multiVendorProduct->getImages()[0]);
+        static::assertSame('alternative text mvp', $multiVendorProduct->getImages()[0]->getAltText());
+    }
+
+    public function testGetMultiVendorProductByIdWithImagesAlt(): void
+    {
+        $service = $this->buildMultiVendorProductService('admin@wizaplace.com', static::VALID_PASSWORD);
+
+        $newMvp = new MultiVendorProduct(
+            [
+                'name' => 'New Test MVP',
+                'code' => 'XXX-xxx-XXX',
+                'supplierReference' => 'AZAZPMPKKD23',
+                'productTemplateType' => 'product',
+                'slug' => 'new-test-mvp',
+                'shortDescription' => 'Nihil est qui quibusdam exercitationem',
+                'description' => 'Nihil est qui quibusdam exercitationem consequatur doloribus sit velit. Ut temporibus est qui et molestiae facilis nisi',
+                'seoTitle' => 'New Test MVP',
+                'seoDescription' => 'New Test MVP Nihil est qui quibusdam exercitationem',
+                'seoKeywords' => 'MVP, PRODUCT, TEST',
+                'status' => 'A',
+                'categoryId' => 3,
+                'freeAttributes' => [
+                    'Free attribute multiple' => [
+                        'Num 1',
+                        'Num 2',
+                        51,
+                    ],
+                    'Free attribute simple' => [
+                        'Bla',
+                    ],
+                ],
+                'attributes' => [],
+            ]
+        );
+
+        $uuid = $service->createMultiVendorProduct($newMvp);
+
+        $image = $this->mockUploadedFile("favicon.png");
+
+        $files = [
+            new MultiVendorProductFile('file', $image->getStream(), $image->getClientFilename()),
+        ];
+
+        $multiVendorProduct = $service->addImageToMultiVendorProduct($uuid, $files, 'alternative text mvp');
+
+        $mvp = $service->getMultiVendorProductById($multiVendorProduct->getId());
+
+        static::assertSame('alternative text mvp', $mvp->getImages()[0]->getAltText());
     }
 
     public function testAddVideoToMultiVendorProductWithHostedFile()