+++ CHANGELOG.md
@@ -6,6 +6,19 @@
 - Added $renderInvoiceOnly parameter to `Wizaplace\SDK\Order\OrderService`
 - Added $renderInvoiceOnly parameter to `Wizaplace\SDK\Vendor\Order\OrderService`
 
+## 1.148.0
+
+### New features
+
+- Added `Wizaplace\SDK\Image\Image::getAltText`
+- Added `Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProduct::getImagesData`
+- Added `Wizaplace\SDK\PIM\Product::getMainImagesData`
+- Added `Wizaplace\SDK\PIM\Product::getAdditionalImagesData`
+- Added `Wizaplace\SDK\PIM\ProductImageUpload::setAltText`
+
+### Changes
+- In the `Wizaplace\SDK\Payment\PaymentService::getPaymentMethods`, it is no more required to be authenticated.
+
 ## 1.147.0
 
 ### New features